requirements.txt
aiogram
openai
deepgram-sdk
python-dotenv